feat: 함께봐여 캘린더에 shareData 적용
